+++ src/rendering/move.c
@@ -6,7 +6,7 @@
 /*   By: owalsh <owalsh@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/14 17:52:35 by owalsh            #+#    #+#             */
-/*   Updated: 2022/11/15 15:52:49 by owalsh           ###   ########.fr       */
+/*   Updated: 2022/11/15 16:12:56 by owalsh           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,14 +33,14 @@ void	render_player(t_game *game, t_img *img, t_coord dest, int color)
 	}
 }
 
-int	is_wall(t_game *game, int dest_x, int dest_y)
+int	is_wall(t_game *game, float dest_x, float dest_y)
 {
-	int	square_x;
-	int	square_y;
+	float	square_x;
+	float	square_y;
 
 	square_y = dest_y * game->map->height / game->mlx->height;
 	square_x = dest_x * game->map->width / game->mlx->width;
-	if (game->map->content[square_y][square_x] == WALL)
+	if (game->map->content[(int)square_y][(int)square_x] == WALL)
 		return (TRUE);
 	return (FALSE);
 }
